\name{AnnotationHub-package}
\alias{AnnotationHub-package}
\docType{package}
\title{Light-weight AnnotationHub 3.0 Client}
\description{
  Client to Bioconductor AnnotationHub 3.0 for discovery and retrieval
  of annotation resources.
}

\author{Martin Morgan \url{mtmorgan@fhcrc.org}}
\keyword{ package }
\examples{
packageDescription("AnnotationHub")
}
