\name{AnnotationHubResource-objects}
\docType{class}

% Classes
\alias{class:AnnotationHubResource}
\alias{AnnotationHubResource-class}

% Constructor

% Accessor-like methods
\alias{getHub}
\alias{getHub,AnnotationHubResource-method}
\alias{hubCache,AnnotationHubResource-method}
\alias{hubUrl,AnnotationHubResource-method}


\title{AnnotationHubResource objects and their related methods and functions}

\description{
    This class is utilized internally. AnnotationHubResources are
    created as an intermediate 'on the fly' process when loading a
    resource. It's accessor methods are used internally.
}

\author{Bioconductor Core Team}

\seealso{\link{AnnotationHub-class}}

\keyword{internal}
\keyword{classes}
\keyword{methods}
