\name{AnnotationHub-package}
\alias{AnnotationHub-package}
\docType{package}
\title{Light-weight AnnotationHub 3.0 Client}
\description{
  Client for discovery and retrieval of Bioconductor annotation resources.
}

\author{Martin Morgan \url{mtmorgan@fhcrc.org}}
\keyword{ package }
\seealso{\code{AnnotationHub-class}}
\examples{
\dontrun{
library(AnnotationHub)
hub = AnnotationHub()
hub
}
}
