\name{AnnotationHubResource-objects}
\docType{class}

% Classes
\alias{class:AnnotationHubResource}
\alias{AnnotationHubResource-class}

% Constructor

% Accessor-like methods
\alias{getHub} 
\alias{getHub,AnnotationHubResource-method} 
\alias{hubCache,AnnotationHubResource-method} 
\alias{hubUrl,AnnotationHubResource-method} 


\title{AnnotationHubResource objects and their related methods and functions}

\description{
    TODO
}

\section{Accessors}{
  In the code snippets below, \code{x} and \code{object} are
  AnnotationHubResource objects.

  \describe{
    \item{}{
      \code{getHub(x)}:
      Gets the AnnotationHub from the hub slot.
    }
    \item{}{
      \code{getCache(x)}:
      Gets the location of the AnnotationHub cache from the AnnotationHub
      object in the hub slot.
    }
    \item{}{
      \code{getUrl(x)}:
      Gets the location of the AnnotationHub production server from the 
      AnnotationHub object in the hub slot.
    }
  }
}

\author{Bioconductor Core Team}

\examples{
    ## TODO
}

\keyword{classes}
\keyword{methods}
